better encapsulating. No odd references to self.xxx when self is not in file
